+++ apps/block_scout_web/lib/block_scout_web/controllers/address_read_contract_controller.ex
@@ -12,7 +12,7 @@ defmodule BlockScoutWeb.AddressReadContractController do
   alias Explorer.ExchangeRates.Token
   alias Indexer.Fetcher.CoinBalanceOnDemand
 
-  import BlockScoutWeb.AddressController, only: [transaction_count: 1, validation_count: 1]
+  import BlockScoutWeb.AddressController, only: [transaction_and_validation_count: 1]
 
   def index(conn, %{"address_id" => address_hash_string}) do
     address_options = [
@@ -27,14 +27,16 @@ defmodule BlockScoutWeb.AddressReadContractController do
 
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.find_contract_address(address_hash, address_options, true) do
+      {transaction_count, validation_count} = transaction_and_validation_count(address_hash)
+
       render(
         conn,
         "index.html",
         address: address,
         coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
-        transaction_count: transaction_count(address_hash),
-        validation_count: validation_count(address_hash)
+        transaction_count: transaction_count,
+        validation_count: validation_count
       )
     else
       :error ->